tp-final/requirements.txt
flask
numpy
pandas
scikit-learn
joblib
mysql-connector-python
schedule
python-dotenv
